requirements.txt
kagglehub
python-dotenv
pyyaml